requirements.txt
Django==2.0.2
pytz==2018.4
djangorestframework==3.8.2